+++ src/restfw/testing/resource_testing.py
@@ -49,6 +49,8 @@ class GetRequestsTester(RequestsTester):
                 assert res.json_body == result
             else:
                 assert res.text == result
+        if result_headers is not None:
+            assert res.headers == result_headers
         head_res = self.web_app.head(self.resource_url, params=params, headers=headers,
                                      exception=exception, status=status)
         assert head_res.headers == res.headers
@@ -76,6 +78,8 @@ class PutRequestsTester(RequestsTester):
                 assert res.json_body == result
             else:
                 assert res.text == result
+        if result_headers is not None:
+            assert res.headers == result_headers
 
 
 @implementer(ISendTestingRequest)
@@ -99,6 +103,8 @@ class PatchRequestsTester(RequestsTester):
                 assert res.json_body == result
             else:
                 assert res.text == result
+        if result_headers is not None:
+            assert res.headers == result_headers
 
 
 @implementer(ISendTestingRequest)
@@ -122,6 +128,8 @@ class PostRequestsTester(RequestsTester):
                 assert res.json_body == result
             else:
                 assert res.text == result
+        if result_headers is not None:
+            assert res.headers == result_headers
 
 
 @implementer(ISendTestingRequest)
@@ -141,6 +149,8 @@ class DeleteRequestsTester(RequestsTester):
                 assert res.json_body == result
             else:
                 assert res.text == result
+        if result_headers is not None:
+            assert res.headers == result_headers
 
 
 def assert_resource(resource_info, web_app):